__init__.py: set waiter result on sucessful read of all chunks.
I left the release call if type is SmartPool just in case, but it
probably needs to be gone too as now all asynchronous exit can be
handled inside connection.py.

The __del__ method creates a task with no certitude of executing,
get rid of it in favor of releasing inside __aexit__ of Connection.