Prevent Connection.close() from being called twice
Connection.close() is occasionally being called more
than once from different code-paths in a most classical
concurrency screw-up.

This was discovered after the persistence of RuntimeErrors
usually coming from readuntil() awaitables proved tricky to
solve.

The commit also ressurects the create_task() method to release
Connection on HttpResponse __del__(), though it binds the task
strongly to a reference inside the Connection object itself,
with the inconvenience that aiosonic must pass a instance
of the running loop - __del__() is special and finnicky
and will complain about a missing loop otherwise.