Ensure aiosonic only close the connection from one place
Issue #473 is still present, just not very often like it used
to be. PR #474 was a first try at getting rid of every
RuntimeError possible. It alleviated the problem but it seems
it is a smarter choice to only really try to forcefully end
the connection from Connection._connect() and only when needed,
i.e. a close to begin a new connection.