Move back to sync code for Connection Release
The work started at #474 aimed to get rid of every RuntimeError.

Now that is understood that we must close the Connection from one
place only in the code to prevent RuntimeError, it appears as
superficial to actually wait on the transport to close and to
make sure a read into the void is done to any trailing buffer.

Doing none of those things still results in functional aiosonic
without RuntimeError's and allow us to go back to sync code for
Connection release and close, thus allowing to relax the need
of pass an instance the Running Loop to HttpResponse.__init__()