Ensure aiosonic only close the connection from one place (#483)
* Ensure aiosonic only close the connection from one place

Issue #473 is still present, just not very often like it used
to be. PR #474 was a first try at getting rid of every
RuntimeError possible. It alleviated the problem but it seems
it is a smarter choice to only really try to forcefully end
the connection from Connection._connect() and only when needed,
i.e. a close to begin a new connection.

* Move back to sync code for Connection Release

The work started at #474 aimed to get rid of every RuntimeError.

Now that is understood that we must close the Connection from one
place only in the code to prevent RuntimeError, it appears as
superficial to actually wait on the transport to close and to
make sure a read into the void is done to any trailing buffer.

Doing none of those things still results in functional aiosonic
without RuntimeError's and allow us to go back to sync code for
Connection release and close, thus allowing to relax the need
of pass an instance the Running Loop to HttpResponse.__init__()

* Abort the Transport unconditionally on close even if closing

According to performance.py it shows an increase of 10% in speed
in relation to aiohttp, consistently. Stats relative to my
machine of course, but the improvement is notable.